requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
<<<<<<< HEAD

charset-normalizer==2.1.0

charset-normalizer==2.1.0

crispy-bootstrap5==0.6

=======
<<<<<<< HEAD
<<<<<<< HEAD
charset-normalizer==2.1.0
=======
>>>>>>> 113443751e7132b0b294df84db685278a2e26bb9
=======
charset-normalizer==2.1.0
>>>>>>> 4a6f097ef043a5e6c9b68297f0b78ca8a9d38ea4
>>>>>>> c5ab4a06a17bd4d9225d92c62369f8da586c1a10
cryptography==37.0.4
dj-database-url==1.0.0
Django==4.0.6
django-on-heroku==1.1.2
gunicorn==20.1.0
django-crispy-forms==1.13.0
h11==0.13.0
idna==3.3
lxml==4.9.1
numpy==1.23.1
outcome==1.2.0
pandas==1.4.3
<<<<<<< HEAD
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1

requests==2.28.1

python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1

=======
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
<<<<<<< HEAD
<<<<<<< HEAD
requests==2.28.1
=======
>>>>>>> 113443751e7132b0b294df84db685278a2e26bb9
=======
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
>>>>>>> 4a6f097ef043a5e6c9b68297f0b78ca8a9d38ea4
>>>>>>> c5ab4a06a17bd4d9225d92c62369f8da586c1a10
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
wsproto==1.1.0
